+++ twilio_integration/twilio_integration/utils.py
@@ -3,7 +3,7 @@ import frappe
 from frappe.utils import get_url
 
 
-def get_public_url(path=None, use_ngrok=True):
+def get_public_url(path: str=None, use_ngrok: bool=True):
 	"""Returns a public accessible url of a site using ngrok.
 	"""
 	if frappe.conf.developer_mode and use_ngrok:
@@ -16,3 +16,13 @@ def get_public_url(path=None, use_ngrok=True):
 		return '/'.join(map(lambda x: x.strip('/'), [domain, path or '']))
 	return get_url(path)
 
+
+def merge_dicts(d1: dict, d2: dict):
+	"""Merge dicts of dictionaries.
+	>>> merge_dicts(
+		{'name1': {'age': 20}, 'name2': {'age': 30}},
+		{'name1': {'phone': '+xxx'}, 'name2': {'phone': '+yyy'}, 'name3': {'phone': '+zzz'}}
+	)
+	... {'name1': {'age': 20, 'phone': '+xxx'}, 'name2': {'age': 30, 'phone': '+yyy'}}
+	"""
+	return {k:{**v, **d2.get(k, {})} for k, v in d1.items()}
